# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Symaira Memory REST API
  description: |
    Local REST HTTP API daemon powering Symaira Memory (symmemory).
    Allows browser extensions, mobile clients, and third-party scripts to read, 
    write, and query persistent semantic memory elements offline.
  version: 0.15.2
servers:
  - url: http://127.0.0.1:8787
    description: Local daemon server
paths:
  /api/status:
    get:
      summary: Get server health and status
      description: Returns the status, version, and server identity.
      responses:
        '200':
          description: Server is active and healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: healthy
                  version:
                    type: string
                    example: 0.15.2
                  server:
                    type: string
                    example: symaira-memory

  /api/search:
    post:
      summary: Semantic vector similarity search
      description: Queries stored memories using cosine similarity over local text embeddings.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - query
              properties:
                query:
                  type: string
                  description: The search term or fact context to match semantically.
                  example: User active project
                scope:
                  type: string
                  description: Optional scope level filter (global, project, agent, user, session).
                  example: project
                limit:
                  type: integer
                  description: Maximum number of search results to return (default 5).
                  example: 3
      responses:
        '200':
          description: Search completed successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Memory'
        '401':
          description: Unauthorized - missing or invalid JWT token

  /api/set:
    post:
      summary: Save a new persistent memory
      description: Stores a new fact in the local SQLite database. Automatically triggers local embedding generation, PII masking, and active project scope parsing.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - content
              properties:
                content:
                  type: string
                  description: The fact content to remember.
                  example: User prefers typescript over python
                scope:
                  type: string
                  description: Scope level (global, project, agent, user, session).
                  example: global
                metadata:
                  type: object
                  additionalProperties:
                    type: string
                  description: Key-value metadata mapping.
                  example:
                    source: extension
      responses:
        '200':
          description: Memory successfully saved
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  id:
                    type: string
                    example: e114ff39-fd89-4efc-8955-2ce5bc67c896
        '401':
          description: Unauthorized

  /api/list:
    get:
      summary: List stored memories
      description: Returns a complete list of stored facts, optionally filtered by scope.
      security:
        - BearerAuth: []
      parameters:
        - name: scope
          in: query
          required: false
          schema:
            type: string
          description: Optional scope filter.
          example: project
      responses:
        '200':
          description: List fetched successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Memory'
        '401':
          description: Unauthorized

  /api/get:
    get:
      summary: Retrieve a single memory by ID
      description: Returns the full memory object for the given UUID.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: string
            format: uuid
          description: Unique memory identifier.
          example: e114ff39-fd89-4efc-8955-2ce5bc67c896
      responses:
        '200':
          description: Memory found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Memory'
        '400':
          description: Bad request - missing id parameter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
        '404':
          description: Memory not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/delete:
    delete:
      summary: Delete a memory by ID
      description: Permanently removes a stored memory. Also accepts POST for clients that cannot send DELETE.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: string
            format: uuid
          description: Unique memory identifier.
          example: e114ff39-fd89-4efc-8955-2ce5bc67c896
      responses:
        '200':
          description: Memory deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  deleted:
                    type: boolean
                    example: true
        '400':
          description: Bad request - missing id parameter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
        '404':
          description: Memory not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      summary: Delete a memory by ID (POST alternative)
      description: Same as DELETE but accepts POST for clients that cannot send DELETE requests.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: string
            format: uuid
          description: Unique memory identifier.
      responses:
        '200':
          description: Memory deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  deleted:
                    type: boolean
                    example: true
        '400':
          description: Bad request - missing id parameter
        '401':
          description: Unauthorized
        '404':
          description: Memory not found

  /api/rules:
    get:
      summary: List behavioral rules
      description: Returns all stored procedural memory rules, optionally filtered by scope.
      security:
        - BearerAuth: []
      parameters:
        - name: scope
          in: query
          required: false
          schema:
            type: string
          description: Optional scope filter.
          example: global
      responses:
        '200':
          description: Rules retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  rules:
                    type: array
                    items:
                      $ref: '#/components/schemas/Rule'
        '401':
          description: Unauthorized

  /api/entities:
    get:
      summary: List all known entities
      description: Returns a list of all known entities (people, projects, organizations) that can be linked to memories.
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Entities retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  entities:
                    type: array
                    items:
                      $ref: '#/components/schemas/Entity'
        '401':
          description: Unauthorized

  /api/token/revoke:
    post:
      summary: Revoke a JWT token
      description: |
        Revokes a JWT token by its jti (or full token value) so it can no
        longer be used to authenticate. Requires a valid bearer token.
        The revocation takes effect in memory immediately and is persisted
        when a revocation store is available; a persistence failure is
        reported as a 500 while the in-memory revocation still applies.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
                  description: Full JWT token value to revoke; its jti claim is extracted automatically.
                jti:
                  type: string
                  description: Raw JWT ID (jti claim) to revoke.
              oneOf:
                - required:
                    - token
                - required:
                    - jti
      responses:
        '200':
          description: Token revoked successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: revoked
                  jti:
                    type: string
                    example: 9f2c4e8a1b3d5c7e9a0b1c2d3e4f5a6b
        '400':
          description: Bad request - missing or invalid token/jti
        '401':
          description: Unauthorized
        '403':
          description: Forbidden (unauthenticated state-changing request blocked by CSRF protection)
        '405':
          description: Method not allowed
        '500':
          description: Token revoked in memory but persistence failed

  /api/sync/changes:
    get:
      summary: Pull memories changed since a timestamp
      description: Returns all memories with updated_at strictly after the supplied RFC 3339 timestamp. If since is missing or invalid, all rows are returned.
      security:
        - BearerAuth: []
      parameters:
        - name: since
          in: query
          required: false
          schema:
            type: string
            format: date-time
          description: RFC 3339 timestamp; only memories updated after this instant are returned.
          example: "2026-06-01T00:00:00Z"
      responses:
        '200':
          description: Changes retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  memories:
                    type: array
                    items:
                      $ref: '#/components/schemas/Memory'
                  server_time:
                    type: string
                    format: date-time
                    description: Current server time in RFC 3339 format.
                    example: "2026-06-10T12:00:00Z"
        '400':
          description: Bad request - invalid since parameter format
        '401':
          description: Unauthorized

  /api/sync/apply:
    post:
      summary: Push memories with last-write-wins conflict resolution
      description: Accepts a batch of memories and upserts each by ID. A row is only overwritten when the incoming updated_at is strictly newer than the stored value.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - memories
              properties:
                memories:
                  type: array
                  items:
                    $ref: '#/components/schemas/Memory'
      responses:
        '200':
          description: Apply completed
          content:
            application/json:
              schema:
                type: object
                properties:
                  applied:
                    type: integer
                    description: Number of memories inserted or updated.
                    example: 3
                  skipped:
                    type: integer
                    description: Number of memories skipped (older or missing ID).
                    example: 1
        '401':
          description: Unauthorized

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Memory:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: e114ff39-fd89-4efc-8955-2ce5bc67c896
        content:
          type: string
          example: User prefers typescript over python
        scope:
          type: string
          example: global
        metadata:
          type: object
          additionalProperties:
            type: string
          example:
            source: extension
            similarity_score: "0.8924"
        embedding:
          type: array
          items:
            type: number
            format: float
          description: Normalized 768-dimensional local text embedding.
        created_at:
          type: string
          format: date-time
          example: "2026-06-02T16:20:23Z"
        updated_at:
          type: string
          format: date-time
          example: "2026-06-02T16:20:23Z"
    Rule:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: b3a7c1f2-9d4e-4a12-b5f6-1e2d3c4b5a69
        content:
          type: string
          example: Always respond in concise bullet points.
        scope:
          type: string
          example: global
        metadata:
          type: object
          additionalProperties:
            type: string
          example:
            source: cli_rule_add
        created_at:
          type: string
          format: date-time
          example: "2026-06-02T16:20:23Z"
    Error:
      type: object
      properties:
        error:
          type: string
          example: not found
    Entity:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: a1b2c3d4-e5f6-7890-abcd-ef1234567890
        name:
          type: string
          example: Alice
        created_at:
          type: string
          format: date-time
          example: "2026-06-02T16:20:23Z"